requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
colour==0.1.5
cryptography==44.0.2
pillow==11.1.0
pycparser==2.22
pyperclip==1.9.0
tkmacosx==1.0.5
